Make sure yesterday is not used as today's match list
I don't really know what goes wrong, but with this commit,
the cache must have been made 15 seconds after midnight.
Hopefully that should avoid situations in which the date was
calculated before midnight, but the cache file was created
right after midnight (busting the assumption that the cache
file's date is the date it was generated on).

Hours are used when calculating the difference, that might
make a difference but I don't know.