Workaround bug in filemtime returning local timestamp
Unix timestamps are supposed to be UTC, but filemtime returns
a Unix timestamp in the local timezone. This caused the oldMatchDate
to be understood as being 2 hours different from the actual
MatchDate (since the local timezone is +0200). Instruct PHP that
the timestamp is in the Europe/Oslo timezone (not a wild guess,
considering the script is made for Norwegian handball).